[Feature] Remove hardcoded limit of media (#11)
Fixes #7

The limit of upload and remote display media was hard coded to 4.

In this PR, I did the following changes:
- Status form now respects instance's upload limit config.
- Remove the limit of images shown on statuses
    - Remote statuses may have more than local instance's limit

Also, I kept the 16:9 aspect ratio of the images. I can change to more complex layouts in future PRs if needed.

This works, but it needs more tests. I don't have how to change the config of my instance, so I had to test forcing duplicates on the post. Any help to validate configuration would be nice.

![image](/attachments/f073ac59-c6d5-4d4c-a01a-17d0c6855094)

Reviewed-on: https://codeberg.org/superseriousbusiness/masto-fe-standalone/pulls/11
Co-authored-by: Thiago 'Jedi' Cerqueira <thiagoa7@gmail.com>
Co-committed-by: Thiago 'Jedi' Cerqueira <thiagoa7@gmail.com>